+++ asf_search/Products/OPERAS1Product.py
@@ -20,6 +20,8 @@ class OPERAS1Product(S1Product):
         'polarization': {
             'path': ['AdditionalAttributes', ('Name', 'POLARIZATION'), 'Values']
         },  # dual polarization is in list rather than a 'VV+VH' style format
+        'tileID': {'path': ['AdditionalAttributes', ('Name', 'MGRS_TILE_ID'), 'Values', 0]},
+        'productVersion': {'path': ['AdditionalAttributes', ('Name', 'PRODUCT_VERSION'), 'Values', 0]},
     }
 
     _subclass_concept_ids = {
@@ -41,6 +43,8 @@ class OPERAS1Product(S1Product):
         'C2803501758-ASF',
         'C3294057315-ASF',
         'C1271830354-ASF',
+        'C1275699124-ASF',
+        'C4090131664-ASF',
     }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
@@ -64,12 +68,16 @@ class OPERAS1Product(S1Product):
         }
 
         if self.properties['processingLevel'] is None:
-            self.properties['processingLevel'] = self.umm_get(self.umm, 'AdditionalAttributes', ('Name', 'PRODUCT_TYPE'), 'Values', 0)
-        
-        # if self.properties['processingLevel'] == 'TROPO-ZENITH':
-        #     west,north,east, south = self.umm['SpatialExtent']['HorizontalSpatialDomain']['Geometry']['BoundingRectangles'][0].values()
+            self.properties['processingLevel'] = self.umm_get(
+                self.umm, 'AdditionalAttributes', ('Name', 'PRODUCT_TYPE'), 'Values', 0)
 
-        #     self.geometry = {'coordinates': [[[west, north], [east,north], [east, south], [west, south], [west, north]]], 'type': 'Polygon'}
+        if self.properties['processingLevel'] == 'DIST-ALERT-S1':
+            west, north, east, south = self.umm['SpatialExtent'][
+                'HorizontalSpatialDomain']['Geometry']['BoundingRectangles'][0].values()
+
+            self.geometry = {'coordinates': [[[west, north], [east, north], [
+                east, south], [west, south], [west, north]]], 'type': 'Polygon'}
+
         if self.properties['processingLevel'] == 'TROPO-ZENITH':
             self.properties['centerLat'] = None
             self.properties['centerLon'] = None